code refactoring, moving functions to utils/duckdb.py
